# src/main/resources/META-INF/openapi.yaml
openapi: 3.0.3
info:
  title: PA POS Microservice
  version: 1.0.0
  description: Public Administration POS of Multi-channel Project
  contact:
    name: Antonio Tarricone
    email: antonio.tarricone@pagopa.it
servers:
  - description: Development Test
    url: https://mil-d-apim.azure-api.net/mil-papos
    x-internal: true
  - description: User Acceptance Test
    url: https://mil-u-apim.azure-api.net/mil-papos
    x-internal: true
...
security:
  - bearerAuth: [ ]
paths:
  /terminals:
    post:
      operationId: createTerminal
      description: Creates a new terminal
      tags: [ terminals, psp ]
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        $ref: '#/components/requestBodies/CreateOrUpdateTerminal'
      responses:
        "201":
          #description: Created
          $ref: '#/components/responses/CreateTerminal'
        "400":
          #description: Bad request
          $ref: '#/components/responses/Error'
        "401":
          #description: Access token is missing or invalid
          $ref: '#/components/responses/Error'
        "403":
          #description: Forbidden
          $ref: '#/components/responses/Error'
        "406":
          #description: Not acceptable. Did you require application/json?
          $ref: '#/components/responses/Error'
        "409":
          #description: Already exists
          $ref: '#/components/responses/Error'
        "415":
          #description: Unsupported media type. Did you provide application/json?
          $ref: '#/components/responses/Error'
        "429":
          #description: Too many requests
          $ref: '#/components/responses/Error'
        "500":
          #description: Server error
          $ref: '#/components/responses/Error'
        default:
          description: Unexpected error
...
components:
...
  schemas:
...
    AccessControlAllowOrigin:
      description: Indicates whether the response can be shared with requesting code from the given origin
      type: string
      pattern: "^[ -~]{1,2048}$"
      minLength: 1
      maxLength: 2048
...
    Enabled:
      description: True if the terminal is enabled
      type: boolean
      example: true
    ErrorCode:
      description: Error code
      type: string
      pattern: "^[A-F0-9]{9}$"
      minLength: 9
      maxLength: 9
      example: "001000005"
    ErrorDescription:
      description: Error description
      type: string
      pattern: "^[ -~]{1,256}$"
      minLength: 1
      maxLength: 256
      example: "Unexpected error from server"
...
    PayeeCode:
      description: Fiscal code of the subject that receives the payment
      type: string
      pattern: "^\\d{11}$"
      minLength: 11
      maxLength: 11
      example: "06534340721"
    PspId:
      description: ID of the POS service provider
      type: string
      pattern: "^[ -~]{1,64}$"
      minLength: 1
      maxLength: 64
      example: "AGID_01"
    RateLimitLimit:
      description: The number of allowed requests in the current period
      type: integer
      format: int32
      minimum: 1
      maximum: 240
    RateLimitReset:
      description: The number of seconds left in the current period
      type: integer
      format: int32
      minimum: 1
      maximum: 60
    RetryAfter:
      description: The number of seconds to wait before allowing a follow-up request
      type: integer
      format: int32
      minimum: 1
      maximum: 240
...
    TerminalId:
      description: ID of the terminal for the PSP
      type: string
      pattern: "^\\d{4,12}$"
      minLength: 4
      maxLength: 12
      example: "34523860"
    TerminalLocation:
      description: URL of the terminal
      type: string
      format: uri
      pattern: "^[ -~]{1,2048}$"
      minLength: 1
      maxLength: 2048
      example: "https://mil-d-apim.azure-api.net/mil-papos/terminals/c7a1b24b0583477292ebdbaa"
...
    UUID:
      description: UUID
      type: string
      pattern: "^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}$"
      minLength: 36
      maxLength: 36
      example: "d0d654e6-97da-4848-b568-99fedccb642b"
...
    CreateOrUpdateTerminal:
      description: Request to create or update a terminal by PSP
      type: object
      additionalProperties: false
      properties:
        pspId:
          $ref: '#/components/schemas/PspId'
        terminalId:
          $ref: '#/components/schemas/TerminalId'
        enabled:
          $ref: '#/components/schemas/Enabled'
        payeeCode:
          $ref: '#/components/schemas/PayeeCode'
      required:
        - pspId
        - terminalId
        - enabled
        - payeeCode
      example:
        pspId: "AGID_01"
        terminalId: "34523860"
        enabled: true
        payeeCode: "06534340721"
...
    Error:
      description: Error details
      type: object
      additionalProperties: false
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        description:
          $ref: '#/components/schemas/ErrorDescription'
      required:
        - code
        - description
      example:
        code: "001000001"
        description: "Generic error"
    Errors:
      description: List of errors
      type: object
      additionalProperties: false
      properties:
        errors:
          type: array
          minItems: 1
          maxItems: 32
          items:
            $ref: '#/components/schemas/Error'
      required:
        - errors
      example:
        errors:
          - code: "00000000A"
            description: "Generic error"
          - code: "001000001"
            description: "Unexpected error from server"
...
  requestBodies:
...
    CreateOrUpdateTerminal:
      description: Request to create a new terminal or update an existing one
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateOrUpdateTerminal'
...
  parameters:
...
    RequestId:
      name: RequestId
      in: header
      description: Request Id that will be logged for troubleshooting purpose
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
...
  responses:
...
    CreateTerminal:
      description: Response of create terminal operation
      headers:
        Access-Control-Allow-Origin:
          description: Indicates whether the response can be shared with requesting code from the given origin
          required: false
          schema:
            $ref: '#/components/schemas/AccessControlAllowOrigin'
        RateLimit-Limit:
          description: The number of allowed requests in the current period
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitLimit'
        RateLimit-Reset:
          description: The number of seconds left in the current period
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitReset'
        Location:
          description: URL of the terminal
          required: true
          schema:
            $ref: '#/components/schemas/TerminalLocation'
...
    Error:
      description: Error response
      headers:
        Access-Control-Allow-Origin:
          description: Indicates whether the response can be shared with requesting code from the given origin
          required: false
          schema:
            $ref: '#/components/schemas/AccessControlAllowOrigin'
        RateLimit-Limit:
          description: The number of allowed requests in the current period
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitLimit'
        RateLimit-Reset:
          description: The number of seconds left in the current period
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitReset'
        Retry-After:
          description: The number of seconds to wait before allowing a follow-up request
          required: false
          schema:
            $ref: '#/components/schemas/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
...
  securitySchemes:
    bearerAuth:
      description: A bearer token in the format of a JWS and conforms to the specifications included in RFC8725
      type: http
      scheme: bearer
      bearerFormat: JWT